# docs/swagger.yaml
definitions:
  api.ApiNameInputJson:
    properties:
      cfg_media:
        type: string
      filepath:
        type: string
      grouptype:
        type: string
      movieid:
        type: integer
      serieid:
        type: integer
    type: object
  api.ApiStructureJson:
    properties:
      configentry:
        type: string
      disabledisallowed:
        type: boolean
      disableruntimecheck:
        type: boolean
      folder:
        type: string
      forceid:
        type: integer
      grouptype:
        type: string
      sourcepathtemplate:
        type: string
      targetpathtemplate:
        type: string
    type: object
  api.apiparse:
    properties:
      config:
        type: string
      name:
        type: string
      path:
        type: string
      quality:
        type: string
      typ:
        type: string
      year:
        type: boolean
    type: object
...
  database.DbmovieJson:
    properties:
      adult:
        type: boolean
      backdrop:
        type: string
      budget:
        type: integer
      createdAt:
        type: string
      facebookID:
        type: string
      freebaseID:
        type: string
      freebaseMID:
        type: string
      genres:
        type: string
      id:
        type: integer
      imdbID:
        type: string
      instagramID:
        type: string
      moviedbID:
        type: integer
      originalLanguage:
        type: string
      originalTitle:
        type: string
      overview:
        type: string
      popularity:
        type: number
      poster:
        type: string
      release_date:
        type: string
      revenue:
        type: integer
      runtime:
        type: integer
      slug:
        type: string
      spokenLanguages:
        type: string
      status:
        type: string
      tagline:
        type: string
      title:
        type: string
      traktID:
        type: integer
      twitterID:
        type: string
      updatedAt:
        type: string
      url:
        type: string
      voteAverage:
        type: number
      voteCount:
        type: integer
      year:
        type: integer
    type: object
...
  database.JobHistoryJson:
    properties:
      createdAt:
        type: string
      ended:
        type: string
      id:
        type: integer
      jobCategory:
        type: string
      jobGroup:
        type: string
      jobType:
        type: string
      started:
        type: string
      updatedAt:
        type: string
    type: object
...
  database.Qualities:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      priority:
        type: integer
      regex:
        type: string
      strings:
        type: string
      type:
        type: integer
      updatedAt:
        type: string
      useRegex:
        type: boolean
    type: object
...
  tasks.Job:
    properties:
      added:
        type: string
      id:
        type: string
      name:
        type: string
      queue:
        type: string
      schedulerId:
        type: string
      started:
        type: string
    type: object
...
info:
  contact: {}
  title: go_media_downloader API
paths:
...
  /api/config/all:
    get:
      consumes:
      - application/json
      description: Get All Config Parameters
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Get Complete Config
      tags:
      - config
  /api/config/clear:
    delete:
      consumes:
      - application/json
      description: Clears the configuration and sets some examples
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Clear Config
      tags:
      - config
  /api/config/delete/{name}:
    delete:
      consumes:
      - application/json
      description: Deletes a configuration
      parameters:
      - description: 'Type Name: ex. quality_SD'
        in: path
        name: name
        required: true
        type: string
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Delete Config
      tags:
      - config
  /api/config/get/{name}:
    get:
      consumes:
      - application/json
      description: Gets a configuration
      parameters:
      - description: 'Type Name: ex. quality_SD'
        in: path
        name: name
        required: true
        type: string
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Get Config
      tags:
      - config
  /api/config/type/{type}:
    get:
      consumes:
      - application/json
      description: List configurations of type
      parameters:
      - description: 'Type Name: ex. quality'
        in: path
        name: type
        required: true
        type: string
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: List Config Type
      tags:
      - config
  /api/config/update/{name}:
    post:
      consumes:
      - application/json
      description: Updates a configuration
      parameters:
      - description: Config
        in: body
        name: config
        required: true
        schema:
          type: object
      - description: 'Type Name: ex. quality_SD'
        in: path
        name: name
        required: true
        type: string
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Update Config
      tags:
      - config
  /api/db/clear/{name}:
    delete:
      consumes:
      - application/json
      description: Clears a DB Table
      parameters:
      - description: Table Name
        in: path
        name: name
        required: true
        type: string
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Clear DB Table
      tags:
      - database
  /api/db/close:
    get:
      consumes:
      - application/json
      description: Closes all database connections
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Close DB
      tags:
      - database
  /api/db/oldjobs:
    delete:
      consumes:
      - application/json
      description: Removes Jobs started over x days ago from db
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      - description: Days ago
        in: query
        name: days
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Clean Old Jobs
      tags:
      - database
  /api/db/vacuum:
    get:
      consumes:
      - application/json
      description: Vacuum database
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Vacuum DB
      tags:
      - database
  /api/fillimdb:
    get:
      consumes:
      - application/json
      description: Downloads IMDB Dataset and creates a new database from it
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Generate IMDB Cache
      tags:
      - general
  /api/movies:
    get:
      consumes:
      - application/json
      description: List Movies
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.DbmovieJson'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: List Movies
      tags:
      - movie
    post:
      consumes:
      - application/json
      description: Updates or creates a movie
      parameters:
      - description: Movie
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/database.DbmovieJson'
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Update Movie (Global)
      tags:
      - movie
...
  /api/naming:
    post:
      consumes:
      - application/json
      description: Test your Naming Convention
      parameters:
      - description: Config
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/api.ApiNameInputJson'
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Name Generation Test
      tags:
      - general
  /api/parse/file:
    post:
      consumes:
      - application/json
      description: Parses a file for testing
      parameters:
      - description: To Parse
        in: body
        name: toparse
        required: true
        schema:
          $ref: '#/definitions/api.apiparse'
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ParseInfo'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Parse a file
      tags:
      - parse
  /api/parse/string:
    post:
      consumes:
      - application/json
      description: Parses a string for testing
      parameters:
      - description: To Parse
        in: body
        name: toparse
        required: true
        schema:
          $ref: '#/definitions/api.apiparse'
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ParseInfo'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Parse a string
      tags:
      - parse
  /api/quality:
    get:
      consumes:
      - application/json
      description: List Qualities with regex filters
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Qualities'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: List Qualities
      tags:
      - quality
    post:
      consumes:
      - application/json
      description: Updates or adds a quality
      parameters:
      - description: Quality
        in: body
        name: quality
        required: true
        schema:
          $ref: '#/definitions/database.Qualities'
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Qualities'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Update Quality
      tags:
      - quality
  /api/quality/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a quality
      parameters:
      - description: Id of Quality to delete
        in: path
        name: id
        required: true
        type: string
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Qualities'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Delete Quality
      tags:
      - quality
  /api/quality/{name}/{config}:
    get:
      consumes:
      - application/json
      description: List allowed qualities and their priorities
      parameters:
      - description: 'Quality Name: ex. quality_SD'
        in: path
        name: name
        required: true
        type: string
      - description: 'Config Name: ex. movie_EN or serie_EN'
        in: path
        name: config
        required: true
        type: string
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/utils.ParseInfo'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: List Quality Priorities
      tags:
      - quality
  /api/queue:
    get:
      consumes:
      - application/json
      description: Lists Queued and Started Jobs (but not finished)
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/tasks.Job'
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Queue
      tags:
      - general
  /api/queue/history:
    get:
      consumes:
      - application/json
      description: Lists Started Jobs and finished but not queued jobs
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order
        type: string
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.JobHistoryJson'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Queue History
      tags:
      - general
  /api/scheduler/list:
    get:
      consumes:
      - application/json
      description: Lists Planned Jobs
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Scheduler Jobs
      tags:
      - scheduler
  /api/scheduler/start:
    get:
      consumes:
      - application/json
      description: Start all Schedulers
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Start Scheduler
      tags:
      - scheduler
  /api/scheduler/stop:
    get:
      consumes:
      - application/json
      description: Stops all Schedulers
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Stop Scheduler
      tags:
      - scheduler
  /api/series:
    get:
      consumes:
      - application/json
      description: List Series
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order
        type: string
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Dbserie'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: List Series
      tags:
      - series
    post:
      consumes:
      - application/json
      description: Updates or creates a series
      parameters:
      - description: Series
        in: body
        name: series
        required: true
        schema:
          $ref: '#/definitions/database.Dbserie'
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Update Series (Global)
      tags:
      - series
...
  /api/structure:
    post:
      consumes:
      - application/json
      description: Structure a single folder
      parameters:
      - description: Config
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/api.ApiStructureJson'
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Structure Single Item
      tags:
      - general
  /api/trakt/authorize:
    get:
      consumes:
      - application/json
      description: List Qualities with regex filters
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Trakt Auhtorize
      tags:
      - general
  /api/trakt/token/{code}:
    get:
      consumes:
      - application/json
      description: Saves Trakt token after Authorization
      parameters:
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Trakt Save Token
      tags:
      - general
  /api/trakt/user/{user}/{list}:
    get:
      consumes:
      - application/json
      description: Get User List
      parameters:
      - description: Trakt Username
        in: path
        name: user
        required: true
        type: string
      - description: List Name
        in: path
        name: list
        required: true
        type: string
      - description: apikey
        in: query
        name: apikey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Trakt Get List (Auth Test)
      tags:
      - general
swagger: "2.0"